style(theGods): 把utils/index.ts的内容移到utils/index.ts
